Marked HandleOnConnected() as virtual
Made both HandleOnConnected() and HandleOnConnectionFailed() as virtual to allow game code to easily handle either of these situations. In a personal project, I needed to have some non-UE4 systems be aware of when the player was connected to Spatial OS; because they were outside of Unreal's API the delegate was useless, This override solved my problem, and I feel like it might be useful for others.